Add TPSL support for MCP
