+++ x10/tools/mcp/private_tools.py
@@ -1,169 +1,16 @@
-import asyncio
-from decimal import Decimal
 from typing import Optional
 
 from mcp.server import FastMCP
 
-from x10.clients.rest import RestApiClient
-from x10.clients.stream import StreamClient
-from x10.config import get_config_by_name
-from x10.core.env_config import EnvConfig
-from x10.core.stark_account import StarkPerpetualAccount
-from x10.errors import ValidationError
-from x10.models.market import MarketModel
-from x10.models.order import OrderSide, OrderType, SelfTradeProtectionLevel, TimeInForce
-from x10.models.orderbook import OrderbookQuantityModel
-from x10.signing.order_object import create_order_object
-from x10.tools.mcp.utils import serialize_tool_result
+from x10.tools.mcp.place_order_tool import register_place_order_tool
+from x10.tools.mcp.utils import create_private_rest_api_client, serialize_tool_result
 from x10.utils.log import get_logger
-from x10.utils.order import get_price_with_slippage
 
 LOGGER = get_logger(__name__)
 
 
-def _create_public_stream_client() -> StreamClient:
-    env_config = EnvConfig.parse()
-    client_config = get_config_by_name(env_config.client_config_name)
-
-    return StreamClient(api_url=client_config.endpoints.stream_url, close_timeout=1)
-
-
-def _create_private_rest_api_client() -> RestApiClient:
-    env_config = EnvConfig.parse()
-    env_config.validate_private_api_credentials()
-    client_config = get_config_by_name(env_config.client_config_name)
-
-    stark_account = StarkPerpetualAccount(
-        api_key=env_config.api_key,
-        public_key=env_config.public_key,
-        private_key=env_config.private_key,
-        vault=env_config.vault_id,
-    )
-
-    return RestApiClient(client_config, stark_account)
-
-
-async def _get_top_of_book(market_name: str) -> tuple[OrderbookQuantityModel, OrderbookQuantityModel] | None:
-    stream_client = _create_public_stream_client()
-
-    LOGGER.debug("Fetching top of book for %s market", market_name)
-
-    async with stream_client.subscribe_to_orderbooks(market_name, depth=1) as orderbook_stream:
-        try:
-            msg = await asyncio.wait_for(orderbook_stream.recv(), timeout=1)
-
-            assert (
-                len(msg.data.bid) == 1 and len(msg.data.ask) == 1
-            ), "Orderbook update does not contain bid or ask data"
-
-            return msg.data.bid[0], msg.data.ask[0]
-        except asyncio.TimeoutError:
-            LOGGER.warn("Timeout while waiting for orderbook update for %s market", market_name)
-        finally:
-            await orderbook_stream.close()
-
-    LOGGER.debug("Falling back to snapshot for %s market", market_name)
-
-    async with stream_client.subscribe_to_orderbooks(market_name) as orderbook_stream:
-        try:
-            msg = await asyncio.wait_for(orderbook_stream.recv(), timeout=1)
-
-            assert (
-                len(msg.data.bid) > 0 and len(msg.data.ask) > 0
-            ), "Orderbook snapshot does not contain bid or ask data"
-
-            return msg.data.bid[0], msg.data.ask[0]
-        except asyncio.TimeoutError:
-            LOGGER.warn("Timeout while waiting for orderbook update for %s market", market_name)
-        finally:
-            await orderbook_stream.close()
-
-    return None
-
-
-async def _get_order_price(
-    *, client: RestApiClient, market: MarketModel, side: OrderSide, order_type: OrderType, price: Decimal | None
-) -> Decimal:
-    if price is not None:
-        return price
-
-    best_bid_and_ask = await _get_top_of_book(market.name)
-
-    if best_bid_and_ask is None:
-        raise ValidationError(f"Failed to fetch top of book for {market.name}")
-
-    best_bid, best_ask = best_bid_and_ask
-
-    return get_price_with_slippage(
-        side=side,
-        price=best_ask.price if side == OrderSide.BUY else best_bid.price,
-        min_price_change=market.trading_config.min_price_change,
-        slippage=client.config.defaults.market_price_slippage,
-    )
-
-
 def register_tools(mcp: FastMCP):
-    @mcp.tool()
-    async def place_order(
-        market_name: str,
-        side: OrderSide,
-        amount_of_synthetic: Decimal,
-        price: Decimal | None = None,
-        order_type: OrderType = OrderType.LIMIT,
-        post_only: bool = False,
-        time_in_force: TimeInForce = TimeInForce.GTT,
-        self_trade_protection_level: SelfTradeProtectionLevel = SelfTradeProtectionLevel.ACCOUNT,
-        external_id: Optional[str] = None,
-        reduce_only: bool = False,
-    ) -> dict:
-        """
-        Place a new order. Requires authentication env vars.
-
-        Args:
-            market_name: Market identifier, e.g. "BTC-USD".
-            side: Order side, one of "BUY" or "SELL".
-            amount_of_synthetic: Order quantity in base asset units.
-            price: Order price. If not provided for MARKET orders, the best bid/ask price will be used.
-            order_type: One of "LIMIT", "MARKET". Defaults to "LIMIT".
-            post_only: If True, the order will be rejected if it would trade immediately.
-            time_in_force: One of "GTT", "IOC", "FOK". Defaults to "GTT".
-            self_trade_protection_level: One of "DISABLED", "ACCOUNT", "CLIENT". Defaults to "ACCOUNT".
-            external_id: Optional client-assigned order ID.
-            reduce_only: If True, the order will only reduce an existing position.
-        """
-
-        if order_type != OrderType.MARKET and not price:
-            raise ValidationError("Price is required for non-MARKET orders")
-
-        async with _create_private_rest_api_client() as client:
-            markets = await client.info.get_markets_dict()
-            market = markets[market_name]
-
-            order_price = await _get_order_price(
-                client=client,
-                market=market,
-                side=side,
-                order_type=order_type,
-                price=price,
-            )
-
-            order = create_order_object(
-                account=client.stark_account,
-                market=market,
-                order_type=order_type,
-                side=side,
-                amount_of_synthetic=amount_of_synthetic,
-                price=order_price,
-                post_only=post_only,
-                time_in_force=time_in_force,
-                reduce_only=reduce_only,
-                order_external_id=external_id,
-                self_trade_protection_level=self_trade_protection_level,
-                starknet_domain=client.config.signing.starknet_domain,
-            )
-
-            result = await client.orders.place_order(order=order)
-            return serialize_tool_result(result.data)
+    register_place_order_tool(mcp)
 
     @mcp.tool()
     async def cancel_order(order_id: int) -> dict:
@@ -174,7 +21,7 @@ def register_tools(mcp: FastMCP):
             order_id: The numeric ID of the order to cancel.
         """
 
-        async with _create_private_rest_api_client() as client:
+        async with create_private_rest_api_client() as client:
             result = await client.orders.cancel_order(order_id=order_id)
             return serialize_tool_result(result.data)
 
@@ -195,7 +42,7 @@ def register_tools(mcp: FastMCP):
             cancel_all: If True, cancel all open orders regardless of other filters.
         """
 
-        async with _create_private_rest_api_client() as client:
+        async with create_private_rest_api_client() as client:
             result = await client.orders.mass_cancel(
                 order_ids=order_ids,
                 external_order_ids=external_order_ids,
@@ -210,7 +57,7 @@ def register_tools(mcp: FastMCP):
         Get account balance. Requires authentication env vars.
         """
 
-        async with _create_private_rest_api_client() as client:
+        async with create_private_rest_api_client() as client:
             result = await client.account.get_balance()
             return serialize_tool_result(result.data)
 
@@ -223,7 +70,7 @@ def register_tools(mcp: FastMCP):
             market_names: Optional list of market names to filter.
         """
 
-        async with _create_private_rest_api_client() as client:
+        async with create_private_rest_api_client() as client:
             result = await client.account.get_positions(market_names=market_names)
             return serialize_tool_result(result.data)
 
@@ -236,6 +83,6 @@ def register_tools(mcp: FastMCP):
             market_names: Optional list of market names to filter.
         """
 
-        async with _create_private_rest_api_client() as client:
+        async with create_private_rest_api_client() as client:
             result = await client.account.get_open_orders(market_names=market_names)
             return serialize_tool_result(result.data)
